Added ability to leave group and easily share groups

+++ src/components/GroupOverview.tsx
@@ -23,18 +23,40 @@ export default class extends Component<Props, State> {
     this._loadGroupData();
   }
 
+  componentWillReceiveProps(): void {
+    this.setState({
+      loading: true,
+      owner: undefined,
+      members: []
+    }, this._loadGroupData);
+  }
+
   _deleteGroup = () => {
     const { group } = this.props;
-    fetch(`/api/v1/groups/${group.id}`, {
-      method: "DELETE"
-    })
-      .then(response => response.json())
-      .then(json => {
-        const { redirect = "/" } = json;
-        window.location.href = redirect;
+    if (confirm("Are your sure you want to delete this group?")) {
+      this.setState({ loading: true });
+      fetch(`/api/v1/groups/${group.id}`, {
+        method: "DELETE"
       })
-      .catch(console.error);
+        .then(response => response.json())
+        .then(json => {
+          const { redirect = "/" } = json;
+          window.location.href = redirect;
+        })
+        .catch(console.error);
+    }
   };
+
+  _leaveGroup = () => {
+    const { group } = this.props;
+    if (confirm("Are your sure you want to leave this group?")) {
+      this.setState({ loading: true });
+      fetch(`/api/v1/groups/${group.id}/leave`, {
+        method: "POST"
+      })
+        .catch(console.error);
+    }
+  }
 
   _loadGroupData = () => {
     Promise.all([this._loadGroupOwner(), this._loadGroupMembers()])
@@ -67,6 +89,17 @@ export default class extends Component<Props, State> {
     });
   };
 
+  _shareGroup = () => {
+    const { group } = this.props;
+    const input = document.createElement('input');
+    document.body.appendChild(input);
+    input.value = `${window.location.host}/api/v1/groups/${group.id}/join`;
+    input.focus();
+    input.select();
+    document.execCommand('copy');
+    document.body.removeChild(input);
+  };
+
   render(): JSX.Element {
     const { group, me } = this.props;
     const { loading, owner } = this.state;
@@ -76,14 +109,26 @@ export default class extends Component<Props, State> {
         <Header>
           {group.name}
         </Header>
-        {owner !== undefined && me.id === owner.id &&
-          <Button
-            color="red"
-            onClick={this._deleteGroup}
-          >
-            Delete
-          </Button>
-        }
+        <Button.Group>
+          <Button onClick={this._shareGroup}>Share</Button>
+          {owner !== undefined && me.id === owner.id &&
+            <Button
+              attached="right"
+              color="red"
+              onClick={this._deleteGroup}
+            >
+              Delete
+            </Button>
+          }
+          {owner !== undefined && me.id !== owner.id &&
+            <Button
+              color="red"
+              onClick={this._leaveGroup}
+            >
+              Leave
+            </Button>
+          }
+        </Button.Group>
       </Segment>
     );
   }
